Refactor fonts

+++ src/text/types.rs
@@ -34,7 +34,7 @@ pub(crate) struct TextDrawData {
 
 pub(crate) trait FontProvider {
     fn get_glyph_id(&self, c: char) -> u32;
-    fn get_line_height(&self, font_size: f32) -> i32;
+    fn get_line_height(&self, font_size: f32) -> f32;
     fn get_advance(&self, from: u32, to: u32, font_size: f32) -> Option<i32>;
     fn get_metric(&self, id: u32, font_size: f32) -> RectPx;
     fn render_glyph(&mut self, id: u32, font_size: f32) -> Option<RectPx>;
